[更新]修复打包后输入系统报错问题 && 修复状态切换时speed过大导致闪现问题

+++ MMO-SERVER/SceneServer/Core/Combat/Attrubute/AttrubuteData.cs
@@ -9,7 +9,7 @@ namespace SceneServer.Core.Combat.Attrubute
     public class AttrubuteData
     {
         //速度
-        public float Speed;
+        public int Speed;
         //最大生命值
         public int MaxHP;
         //最大魔法值
